network: enable ipv6 on interfaces where ipv4 assignment exists
unfortunately we have to enable dhcpv6 on a seperate subinterface,
even though only router advertisements and solicitations are needed.

i also tested accept_ra=2 in sysctl for the interface with a positive
result, but this is being resetted by uci everytime, therefore do it
the "proper openwrt" way.

partly fixes #170