get rid of compat_version checking
this involves hardcoding the resulting compat_version in the image metadata
to 1.0 and also hardcoding it within the config. We dont need the compat_version
check because we ship the config within the image.

If the device you are flashing is running at a higher compat_version
you likely have to use sysupgrade -n in order to fix it again