github: disable linting rules about comments
These two linters make sense for functional comments,
e.g. when API documentation is generated from comments
and a certain format and syntax needs to be enforced.

Our comments are purely for conveying free-form information.
We don't have a need for any specific formatting of comments.

The false-positive rate of the comment rules is obviously
very high and leads to frustration that can be avoided at no cost.
It's also not good practice to regularly ignore linter errors.

The best way forward is to not lint comments.