wikiupdater: use locations instead of nodes
Add ability to follow a redirect if a wikiarticle is not named Berlin:Standorte:$LOCATION.
It expects at least a redirect.

This change is necessary as currently one article can have several nodes because it adds every node in a certain radius. This means that wikiupdater cant know if it has the right article. Using defined names for articles solves this problem.