inventory: back to acceptable performance
We have ~300 hosts in ~100 locations and we were calling `yq` multiple times
for every single one of them. That was ~700 python interpreter invocations.

Now we only call it once to convert all YAML to JSON in one go,
and let one `jq` script do all the merging. It's less code and much faster.

- Before: `ansible-inventory --list` 0m35,404s
- After: `ansible-inventory --list`  0m4,649s
- Without Ansible: `./inventory/base_inventory --list`  0m0,543s