templates: deactivate stderr of collectd if it monitors airos_6
When collectd wants to get the Names of Stations via snmp connnected
to a device running AirOS 6 that have none connected it reports a
"NoSuchName" error to logd resulting in a lots of logspamming. At
chris-core it was 60 per minute.

This Script will change the option "procd_set_param stderr  1" to "0"
in the file /etc/init.d/collectd to deactivate logging of all errors.
The monitoring will still work. I prefer not having logs over logspamming
as the logfiles will get useless otherwise. It is still possible to
change this option back to 1 until next boot to get a better insight
when debugging collectd-issues.

This fixes #766