snmp_profiles: switch TypeINstanceOID from ifIndex to port labels
Use the Name or Label of the Interface. Depending on the configuration
this might not be unique, however it helps inunderstanding the produced
data as there is otherwise no easy way of knowing which Index belongs
to which Interface. E.g. the first Interfaces usually are used for the
physical ports, but e.g. tplink has internal Interfaces like loopback
first.