network: disable EEE on Mediatek-based devices
EEE is Energy Efficient Ethernet and the MT7530 switch chip in many
Mediatek-based devices has issues with it. The problems seem to be
in either the DSA driver or the OpenWrt integration.

For now we'll disable EEE to prevent the link flapping issue.

    [3549169.529898] mt7530-mdio mdio-bus:1f lan: Link is Down
    [3549172.238591] mt7530-mdio mdio-bus:1f lan: Link is Up - 1Gbps/Full - flow control rx/tx
    [3552592.237232] mt7530-mdio mdio-bus:1f lan: Link is Down
    [3552595.115661] mt7530-mdio mdio-bus:1f lan: Link is Up - 1Gbps/Full - flow control rx/tx

See https://github.com/openwrt/openwrt/issues/17351
and https://stijn.tintel.eu/blog/2024/12/03/disable-eee-openwrt-dsa-interface/