network: fix stray ip6addr option
This piece was outside of the corresponding `network` UCI block.
In certain cases this `ip6addr` option was printed alone.
This means it was parsed as if it belonged to the `network` block
that printed right before.

No content change, just moving it up into the if-endif-block that
was being closed right above.