collectd: set MaxPacketSize
Without this option collectd created packages that where to big resulting in the error
"network plugin: sendto failed: Message too large. Closing sending socket."

This size limits the package to the same value as the mtu for tunnels.
We could optimize the Value in the future, but for now it works.

This problem affected newyorck-core