Fix smallies (#1302)
* network: fix stray ip6addr option

This piece was outside of the corresponding `network` UCI block.
In certain cases this `ip6addr` option was printed alone.
This means it was parsed as if it belonged to the `network` block
that printed right before.

No content change, just moving it up into the if-endif-block that
was being closed right above.

* merge_vars: fix shared-prefix conflicts

Example: the suffix "files__to_merge" also matches
"wireless_profiles__to_merge" which caused funny issues.

Let's also match on the leading underscores to prevent that.

* treewide: consistent chmod file mode syntax