Revert "FBLipke: Add Location"
The location breaks our CI, so remove it. We are happy if the person
uses our freifunk berlin ipv4 space, or is contributing to our templates
to fit their usecase. Please fix the location and do another PR adding
the location again. Sadly, the location will now not experience the
gateway fast firewall bypasses.

This reverts commit ae482b10d694438ad660875d356e747da82b26ed.