fix: Deploy community ssh-keys only if community is explicitly set
Only deploy ssh keys ment for maintenance by the community if the
community attribute is explicitly set. Otherwise an undefined community
attribute deploys ssh keys silently and without warning on private
installations. This new behavior seems to be inline with other uses
of the community attribute in the code.

This change may break standard behavior and requires us to add another
commit that adds the community attribute explicitly to all community
maintained installations.